requirements.txt
diffusers==0.17.0
gymnasium[atari,accept-rom-license]==0.29.1
hydra-core==1.3
matplotlib
opencv-python
pygame
torch==2.0.1
tqdm
wandb
